feat(user): add infinite loading to user sales table
